+++ araucaria/plot/template.py
@@ -200,9 +200,9 @@ def fig_xas_template(panels: str='xx', fig_pars: FigPars=None,
             if k == 0:
                 ax.set_ylabel(r'$\chi(k)$')
             elif k == 1:
-                ax.set_ylabel(r'$k\chi(k)$ [$\AA$]')
+                ax.set_ylabel(r'$k \, \chi(k)$ [$\AA$]')
             else:
-                ax.set_ylabel(r'$k^%i\chi(k)$ [$\AA^{%i}$]' % (k, k) )
+                ax.set_ylabel(r'$k^%i \, \chi(k)$ [$\AA^{%i}$]' % (k, k) )
             keys = ['k_range', 'chi_range', 'k_ticks', 'chi_ticks']
 
         # FT-EXAFS axis
@@ -217,9 +217,9 @@ def fig_xas_template(panels: str='xx', fig_pars: FigPars=None,
             if k == 0:
                 ax.set_ylabel(r'$\chi(q)$')
             elif k == 1:
-                ax.set_ylabel(r'$q\chi(q)$ [$\AA$]')
+                ax.set_ylabel(r'$q \, \chi(q)$ [$\AA$]')
             else:
-                ax.set_ylabel(r'$q^%i\chi(q)$ [$\AA^{%i}$]' % (k, k) )
+                ax.set_ylabel(r'$q^%i \, \chi(q)$ [$\AA^{%i}$]' % (k, k) )
             keys = ['q_range', 'chi_range', 'q_ticks', 'chi_ticks']   
         
         # unnasigned axis