+++ docs/source/changelog.rst
@@ -3,9 +3,11 @@ Changelog
 
 ver 0.1.12
 ----------
-- :mod:`~.fit.feffit` submodule added: support for feffit on EXAFS spectra (:mod:`~araucaria.fit` module).
+- :mod:`~.fit.feffit` submodule added: support for Feffit on EXAFS spectra (:mod:`~araucaria.fit` module).
 - binning functions moved to (:mod:`~araucaria.xas.xasutils` module).
 - added :func:`~.utils.check_dictkeys` function (:mod:`~araucaria.utils` module).
+- added :func:`~.utils.check_maxminval` and :func:`~.utils.check_minmaxval` functions (:mod:`~araucaria.utils` module).
+- added :func:`~.utils.count_decimals` function (:mod:`~araucaria.utils` module).
 - updated labels for EXAFS plots in :func:`~.plot.template.fig_xas_template` function (:mod:`~.plot` module).